+++ neo/sys/sys_localuser.h
@@ -74,12 +74,12 @@ public:
 
 	void WriteToMsg( idBitMsg& msg )
 	{
-		msg.WriteLong( handle );
+		msg.WriteInt( handle );
 	}
 
 	void ReadFromMsg( const idBitMsg& msg )
 	{
-		handle = msg.ReadLong();
+		handle = msg.ReadInt();
 	}
 
 	void Serialize( idSerializer& ser );
